requirements.txt
pandas==1.5.3
matplotlib==3.7.1
pyodbc==5.0.1
python-dotenv==1.0.0
